extract components into seperate files, remove unneeded styled divs, fix style when collapsed
